requirements.txt
Flask
Flask-Cors
Flask-JWT-Extended
Flask-Migrate
SQLAlchemy
PyJWT
python-dotenv
bcrypt
spacy
nltk
scikit-learn
celery
pyppeteer
selenium
python-docx
PyPDF2
gunicorn

